ASoC: sun4i-spdif: Add support for separate RX/TX clocks
On older variants of the hardware, the RX and TX blocks share a single
module clock, named "spdif" in the DT binding. The D1 variant has
separate RX and TX clocks, so the TX module clock is named "tx" in the
binding. To support this, supply the clock name in the quirks structure.

Since the driver supports only TX, only the TX clock name is needed.

Signed-off-by: Samuel Holland <samuel@sholland.org>