riscv, kexec_file: Correct kernel size and alignment
TODO: CHECK IF STILL APPLICABLE (from older patch set)

When adding the initrd to the kexec image, incorrect kernel size is
used. The kernel_len, which is used is the file size, and does not
correspond to the actual memory that the kernel occupies.

Further, when CONFIG_STRICT_KERNEL_RWX the kernel memory region needs
to be aligned.

An example, to show the problem:
  [  172.740749] kexec_image: The entry point of kernel at 0x80200000
  [  172.743812] kexec_image: Loaded initrd at 0x815da000
  [  172.745321] kexec_image: Loaded device tree at 0x47fffd000
  [  173.071537] kexec_core: Starting new kernel
  ...
  [    0.000000] memblock_reserve: [0x0000000080200000-0x00000000815fffff] paging_init+0xda/0x3ae
  ...
  [    0.000000] INITRD: 0x815da000+0x10e8f000 overlaps in-use memory region
  [    0.000000]  - disabling initrd

This patch fixes both issues.

Signed-off-by: Björn Töpel <bjorn@kernel.org>