drm/sun4i: Allow VI layers to be primary planes
D1's mixer 1 has no UI layers, only a VI layer. So that mixer can only
be used if the primary plane comes from the VI layer. Add the code to
handle this case, and set up the mixer's global registers in the same
way as when a UI layer's plane is primary.

Signed-off-by: Samuel Holland <samuel@sholland.org>