riscv: dts: starfive: correct number of external interrupts
The PLIC integrated on the Vic_U7_Core integrated on the StarFive JH7100
SoC actually supports 133 external interrupts.  127 of these are exposed
to the outside world; the remainder are used by other devices that are
part of the core-complex such as the L2 cache controller. But all 133
interrupts are external interrupts as far as the PLIC is concerned.
Fix the property so that the driver can manage these additional external
interrupts, which is important since the interrupts for the L2 cache
controller are enabled by default.

Fixes: ec85362fb121 ("RISC-V: Add initial StarFive JH7100 device tree")
Signed-off-by: Mark Kettenis <kettenis@openbsd.org>
Link: https://lore.kernel.org/r/20220606162924.71418-1-kettenis@openbsd.org